Refactoring for a Python 3 only version
