+++ bigml/bigmlconnection.py
@@ -43,7 +43,7 @@ import bigml.constants as c
 
 from bigml.util import (check_dir,
                         maybe_save, get_exponential_wait)
-from bigml.util import DEFAULT_LOCALE, PY3
+from bigml.util import DEFAULT_LOCALE
 from bigml.domain import Domain
 from bigml.domain import DEFAULT_DOMAIN, BIGML_PROTOCOL
 
@@ -120,8 +120,6 @@ def json_load(content):
 
     """
     args = [content.decode('utf-8')]
-    if not PY3:
-        args.append('utf-8')
     return json.loads(*args)
