+++ bigml/tests/test_44_compare_predictions.py
@@ -20,7 +20,6 @@
 
 """
 from .world import world, setup_module, teardown_module, show_doc
-from bigml.util import PY3
 from . import create_source_steps as source_create
 from . import create_dataset_steps as dataset_create
 from . import create_association_steps as association_create
